% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scouting_utils.R
\name{dv_scout_from_console}
\alias{dv_scout_from_console}
\title{Enter scout codes from the console}
\usage{
dv_scout_from_console(
  x,
  prompt = "SCOUT> ",
  compound_table = ov_default_compound_table(),
  default_scouting_table = ov_default_scouting_table()
)
}
\arguments{
\item{x}{datavolley: a datavolley object as returned by \code{\link[=dv_create]{dv_create()}}}

\item{prompt}{string: the prompt to show}

\item{compound_table}{tibble: the table of default compound codes}

\item{default_scouting_table}{tibble: the table of scouting defaults (skill type and evaluation)}
}
\value{
A modified version of \code{x}, with rows added to the plays2 component
}
\description{
Probably only useful for testing.
}
