% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_scouter.R
\name{ov_app_styling}
\alias{ov_app_styling}
\title{Styling to apply to the app}
\usage{
ov_app_styling()
}
\value{
A named list of styling parameters
}
\description{
Colours can be either hex strings or names, but if using names they must be recognized both by R and as CSS colour names.
}
\details{
\itemize{
\item \code{review_pane_width} is expressed as a percentage of the browser window width
}
}
