% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{ov_default_key_remapping}
\alias{ov_default_key_remapping}
\title{Default keyboard remapping for ov_scouter}
\usage{
ov_default_key_remapping(scout_mode = "click")
}
\arguments{
\item{scout_mode}{string: either "click" for the guided point-and-click scouting interface, or "type" for the typing-based interface. Currently remapping has no effect with "click"}
}
\value{
A named list
}
\description{
These key remappings only apply to input into the scouting bar (when the scouter has been started with \code{scout_mode = "type"}).
}
