% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{ov_default_scouting_table}
\alias{ov_default_scouting_table}
\title{Default scouting (type and evaluation for each skill) table}
\usage{
ov_default_scouting_table()
}
\value{
A tibble
}
\description{
Default scouting (type and evaluation for each skill) table
}
