% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{ov_default_shortcuts}
\alias{ov_default_shortcuts}
\alias{ov_default_playstable_shortcuts}
\title{Default keyboard shortcuts for ov_scouter}
\usage{
ov_default_shortcuts(scout_mode = "click")

ov_default_playstable_shortcuts(scout_mode = "click")
}
\arguments{
\item{scout_mode}{string: either "click" for the guided point-and-click scouting interface, or "type" for the typing-based interface}
}
\value{
A named list
}
\description{
Default keyboard shortcuts for ov_scouter
}
