% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{ov_default_winning_symbols}
\alias{ov_default_winning_symbols}
\title{Default winning symbols table}
\usage{
ov_default_winning_symbols()
}
\value{
A tibble
}
\description{
Default winning symbols table
}
