% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_scouter.R
\name{ov_scouter_options}
\alias{ov_scouter_options}
\title{Scouting options}
\usage{
ov_scouter_options(nblockers = TRUE)
}
\arguments{
\item{nblockers}{logical: scout the number of blockers on each attack?}
}
\value{
A named list
}
\description{
Scouting options
}
