% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{ov_shiny_video_sync}
\alias{ov_shiny_video_sync}
\title{Launch a Shiny app for video synchronisation and scout editing}
\usage{
ov_shiny_video_sync(
  dvw,
  video_file = NULL,
  launch_browser = TRUE,
  prompt_for_files = interactive(),
  ...
)
}
\arguments{
\item{dvw}{string or datavolley: either the path to a dvw file (which will be read by \code{\link[datavolley]{read_dv}}) or a datavolley object as returned by that function. Passing the file name (not the datavolley object) is required if any extra arguments are passed via \code{...}}

\item{video_file}{string: optionally, the path to the video file. If not supplied (or \code{NULL}) the video file specified in the dvw file will be used. Provide \code{video_file = NA} to run the app without a video file}

\item{launch_browser}{logical: if \code{TRUE}, launch the app in the system's default web browser (passed to \code{\link[shiny]{runApp}}'s \code{launch.browser} parameter)}

\item{prompt_for_files}{logical: if \code{dvw} was not specified, prompt the user to select the dvw file}

\item{...}{: extra parameters passed to \code{\link[datavolley]{read_dv}} (if \code{dvw} is a provided as a string) and/or to the shiny server and UI functions}
}
\description{
Launch a Shiny app for video synchronisation and scout editing
}
\seealso{
\code{\link[datavolley]{read_dv}}
}
