% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substitutions_handler.R
\name{rotations}
\alias{rotations}
\title{Get team rotations point by point}
\usage{
rotations(x, team, start_point_id, set_number, new_rotation = NULL)
}
\arguments{
\item{x}{datavolley or data.frame: a datavolley object as returned by \code{read_dv}, or the plays component of that object}

\item{team}{string: team name}

\item{new_rotation}{vector: vector of player number, positionned from 1 to 6.}

\item{point_ids}{vector: vector of pointids for which to return the rotation}
}
\value{
list of 2 data.frames. Current data.frame, without changes, and updated data.frame, with new player rotation.
}
\description{
Get team rotations point by point
}
\examples{
\dontrun{
  x <- read_dv(dv_example_file(), insert_technical_timeouts=FALSE)
  new_x <- rotations(x, start_point_id = 25, new_rotation = c(9,6,15,4,12,7))
  
  toreplace <- x$plays[x$plays$point_id \%in\% new_x$new_rotation$point_id, colnames(new_x$new_rotation)]
  new_xx <- dplyr::left_join(dplyr::select(toreplace, point_id), new_x$new_rotation)
  
  x$plays[x$plays$point_id \%in\% new_x$new_rotation$point_id, colnames(new_x$new_rotation)] <- new_xx
  
}
}
\seealso{
\code{\link{read_dv}}
}
