% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{vwModalDialog}
\alias{vwModalDialog}
\title{Variable width modal dialog}
\usage{
vwModalDialog(..., width = 90)
}
\arguments{
\item{...}{: as for \code{\link[shiny:modalDialog]{shiny::modalDialog()}}}

\item{width}{numeric: percentage of viewport width}
}
\value{
As for \code{\link[shiny:modalDialog]{shiny::modalDialog()}}
}
\description{
Variable width modal dialog
}
\examples{
\dontrun{
  showModal(vwModalDialog(title = "Wide dialog", "blah", width = 90))
}

}
