% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{ov_default_click_shortcuts}
\alias{ov_default_click_shortcuts}
\alias{ov_default_type_shortcuts}
\alias{ov_default_playstable_shortcuts}
\title{Default keyboard shortcuts for ov_scouter}
\usage{
ov_default_click_shortcuts()

ov_default_type_shortcuts()

ov_default_playstable_shortcuts()
}
\value{
A named list
}
\description{
\code{ov_default_click_shortcuts} apply when \verb{using scout_mode = "click"}, and \code{ov_default_type_shortcuts} apply when \verb{using scout_mode = "type"}
}
\details{
Shortcuts should be defined in terms of the printable representation of the key (e.g. "a", "$", "H", "Escape", "Enter"). See \url{https://developer.mozilla.org/en-US/docs/Web/API/UI_Events/Keyboard_event_key_values} for guidance. Shortcuts can optionally use modifier keys: "Ctrl-x" means pressing the control key and x simultaneously; similarly Alt-x, Meta-x, Shift-x.
Be aware that some keys are hard-coded for specific functionality and might cause problems if you use them as shortcuts (e.g. "Enter", "Tab"), and some keys have browser-level or operating-system-level handling that cannot be overridden.
}
