% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{ov_default_click_shortcuts}
\alias{ov_default_click_shortcuts}
\alias{ov_default_type_shortcuts}
\alias{ov_default_playstable_shortcuts}
\title{Default keyboard shortcuts for ov_scouter}
\usage{
ov_default_click_shortcuts()

ov_default_type_shortcuts()

ov_default_playstable_shortcuts()
}
\value{
A named list
}
\description{
\code{ov_default_click_shortcuts} apply when \verb{using scout_mode = "click"}, and \code{ov_default_type_shortcuts} apply when \verb{using scout_mode = "type"}
}
