% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{ov_default_setter_calls_table}
\alias{ov_default_setter_calls_table}
\title{Default setter calls table}
\usage{
ov_default_setter_calls_table()
}
\value{
A tibble
}
\description{
Default setter calls table
}
