% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{ov_default_shortcuts}
\alias{ov_default_shortcuts}
\title{Default keyboard shortcuts for ov_scouter}
\usage{
ov_default_shortcuts()
}
\value{
A named list
}
\description{
Default keyboard shortcuts for ov_scouter
}
