% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_scouter.R
\name{ov_scouter}
\alias{ov_scouter}
\title{Launch a Shiny app for scouting}
\usage{
ov_scouter(
  dvw,
  video_file,
  court_ref,
  season_dir,
  scoreboard = TRUE,
  ball_path = FALSE,
  review_pane = TRUE,
  playlist_display_option = "dv_codes",
  scouting_options = ov_scouter_options(),
  default_scouting_table = ov_default_scouting_table(),
  compound_table = ov_default_compound_table(),
  launch_browser = TRUE,
  prompt_for_files = interactive(),
  ...
)
}
\arguments{
\item{dvw}{string or datavolley: either the path to a dvw or ovs file or a datavolley object (e.g. as returned by \code{\link[=dv_create]{dv_create()}}. Passing the file name (not the datavolley object) is required if any extra arguments are passed via \code{...}. \code{dvw} can also be an object as saved by \code{ov_scouter()} in ovs format. If \code{dvw} is "demo", the app will be started with a demonstration data set}

\item{video_file}{string: optionally, the path to the video file. If not supplied (or \code{NULL}) the video file specified in the dvw file will be used. Provide \code{video_file = NA} to run the app without a video file}

\item{court_ref}{data.frame or string: data.frame with the court reference (as returned by \code{\link[ovideo:ov_shiny_court_ref]{ovideo::ov_shiny_court_ref()}}) or the path to the rds file containing the output from this}

\item{season_dir}{string: optional path to a directory with other dvw/ovs files from this season}

\item{scoreboard}{logical: if \code{TRUE}, show a scoreboard in the top-right of the video pane}

\item{ball_path}{logical: if \code{TRUE}, show the ball path on the court inset diagram. Note that this will slow the app down slightly}

\item{review_pane}{logical: if \code{TRUE}, entry popups will be accompanied by a small video pane that shows a loop of the video of the action in question}

\item{scouting_options}{list: a named list with entries as per \code{\link[=ov_scouter_options]{ov_scouter_options()}}}

\item{default_scouting_table}{tibble: the table of scouting defaults (skill type and evaluation)}

\item{compound_table}{tibble: the table of compound codes}

\item{launch_browser}{logical: if \code{TRUE}, launch the app in the system's default web browser (passed to \code{\link[shiny:runApp]{shiny::runApp()}}'s \code{launch.browser} parameter)}

\item{prompt_for_files}{logical: if \code{dvw} was not specified, prompt the user to select the dvw file}

\item{...}{: extra parameters passed to \code{\link[datavolley:read_dv]{datavolley::dv_read()}} (if \code{dvw} is a provided as a string) and/or to the shiny server and UI functions}
}
\description{
Launch a Shiny app for scouting
}
\examples{
\dontrun{
  ov_scouter("demo")
}

}
