% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_scouter.R
\name{ov_scouter}
\alias{ov_scouter}
\title{Launch a Shiny app for scouting}
\usage{
ov_scouter(
  dvw,
  video_file,
  court_ref,
  season_dir,
  auto_save_dir,
  scoreboard = TRUE,
  ball_path = FALSE,
  review_pane = TRUE,
  playlist_display_option = "dv_codes",
  scouting_options = ov_scouting_options(),
  app_styling = ov_app_styling(),
  shortcuts = ov_default_shortcuts(),
  scout_name = "",
  show_courtref = FALSE,
  host,
  launch_browser = TRUE,
  prompt_for_files = interactive(),
  ...
)
}
\arguments{
\item{dvw}{string or datavolley: either the path to a dvw or ovs file or a datavolley object (e.g. as returned by \code{\link[=dv_create]{dv_create()}}. Passing the file name (not the datavolley object) is required if any extra arguments are passed via \code{...}. \code{dvw} can also be an object as saved by \code{ov_scouter()} in ovs format. If \code{dvw} is "demo", the app will be started with a demonstration data set}

\item{video_file}{string: optionally, the path to the video file. If not supplied (or \code{NULL}) the video file specified in the dvw file will be used. \code{video_file} can also be a URL (including a YouTube URL or video ID)}

\item{court_ref}{data.frame or string: data.frame with the court reference (as returned by \code{\link[ovideo:ov_shiny_court_ref]{ovideo::ov_shiny_court_ref()}}) or the path to the rds file containing the output from this}

\item{season_dir}{string: optional path to a directory with other dvw/ovs files from this season}

\item{auto_save_dir}{string: optional path to a directory where the dvw will be saved automatically after each rally}

\item{scoreboard}{logical: if \code{TRUE}, show a scoreboard in the top-right of the video pane}

\item{ball_path}{logical: if \code{TRUE}, show the ball path on the court inset diagram. Note that this will slow the app down slightly}

\item{review_pane}{logical: if \code{TRUE}, entry popups will be accompanied by a small video pane that shows a loop of the video of the action in question}

\item{playlist_display_option}{string: what to show in the plays table? Either "dv_codes" (scouted codes) or "commentary" (a plain-language interpretation of the touches)}

\item{scouting_options}{list: a named list with entries as per \code{\link[=ov_scouting_options]{ov_scouting_options()}}. See Details, below}

\item{app_styling}{list: named list of styling options, as returned by \code{\link[=ov_app_styling]{ov_app_styling()}}}

\item{shortcuts}{list: named list of keyboard shortcuts, as returned by \code{\link[=ov_default_shortcuts]{ov_default_shortcuts()}}}

\item{scout_name}{string: the name of the scout (your name)}

\item{show_courtref}{logical: if \code{TRUE}, show the court reference lines overlaid on the video}

\item{host}{string: the IP address of this machine. Only required if you intend to connect to the app from a different machine (in which case use \code{ov_scouter(..., host = "www.xxx.yyy.zzz", launch_browser = FALSE)}, where www.xxx.yyy.zzz is the IP address of this machine, i.e. the machine running the app)}

\item{launch_browser}{logical: if \code{TRUE}, launch the app in the system's default web browser (passed to \code{\link[shiny:runApp]{shiny::runApp()}}'s \code{launch.browser} parameter)}

\item{prompt_for_files}{logical: if \code{dvw} was not specified, prompt the user to select the dvw file}

\item{...}{: extra parameters passed to \code{\link[datavolley:read_dv]{datavolley::dv_read()}} (if \code{dvw} is a provided as a string) and/or to the shiny server and UI functions}
}
\description{
Launch a Shiny app for scouting
}
\details{
A note on \code{scouting_options}

If a *.ovs file (i.e. a partially-scouted file, that was previously scouted using this app) has been provided in the \code{dvw} argument, then it will contain the scouting options used during the previous scouting session. Those options will be re-used EXCEPT if \code{scouting_options} are also provided here. Any scouting options provided here via the \code{scouting_options} argument will override options saved in the .ovs file. Thus, it is recommended that \code{scouting_options} not be provided here along with a .ovs file unless absolutely necessary. If necessary, only the specific, relevant elements of the \code{scouting_options} list should be provided. Note that *.dvw files do not contain saved options, only .ovs files that were scouted with this app.
}
\examples{
\dontrun{
  ov_scouter("demo")
}

}
