% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_scouter.R
\name{ov_scouter_options}
\alias{ov_scouter_options}
\title{Scouting options}
\usage{
ov_scouter_options(
  nblockers = TRUE,
  default_nblockers = NA,
  transition_sets = FALSE,
  team_system = "SHM3"
)
}
\arguments{
\item{nblockers}{logical: scout the number of blockers on each attack?}

\item{default_nblockers}{integer: if \code{nblockers} is TRUE, what number of blockers should we default to? If \code{NA}, no default}

\item{transition_sets}{logical: scout sets in transition? If \code{FALSE}, just the endpoint of each attack (i.e. the dig) and the subsequent counter-attack are scouted}

\item{team_system}{string: the assumed system that teams are using to assign e.g. passing and hitting responsibilities
\itemize{
\item "SHM3" - a setter-hitter-middle rotation, with 3 passers (the libero and two outside hitters)
}}
}
\value{
A named list
}
\description{
Scouting options
}
