% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_scouter.R
\name{ov_scouter_options}
\alias{ov_scouter_options}
\title{Scouting options}
\usage{
ov_scouter_options(
  attack_end = "actual",
  nblockers = TRUE,
  default_nblockers = NA,
  transition_sets = FALSE,
  team_system = "SHM3",
  setter_dump_code = "PP",
  second_ball_attack_code = "P2",
  overpass_attack_code = "PR"
)
}
\arguments{
\item{attack_end}{string: "actual" or "intended" the end coordinate of an attack is the actual end location, or the intended one. The actual might differ from the intended if there is a block touch. If "actual", and a block touch is recorded, then the end location of the attack will not be used for the dig location (the dig location will be missing)}

\item{nblockers}{logical: scout the number of blockers on each attack?}

\item{default_nblockers}{integer: if \code{nblockers} is TRUE, what number of blockers should we default to? If \code{NA}, no default}

\item{transition_sets}{logical: scout sets in transition? If \code{FALSE}, just the endpoint of each attack (i.e. the dig) and the subsequent counter-attack are scouted}

\item{team_system}{string: the assumed system that teams are using to assign e.g. passing and hitting responsibilities
\itemize{
\item "SHM3" - a setter-hitter-middle rotation, with 3 passers (the libero and two outside hitters)
}}

\item{setter_dump_code}{string: the attack combination code for a setter dump}

\item{second_ball_attack_code}{string: the attack combination code for a second-ball attack}

\item{overpass_attack_code}{string: the attack combination code for an attack on an overpass}
}
\value{
A named list
}
\description{
Scouting options
}
