% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_scouter.R
\name{ov_scouter_options}
\alias{ov_scouter_options}
\title{Scouting options}
\usage{
ov_scouter_options(nblockers = TRUE, transition_sets = FALSE)
}
\arguments{
\item{nblockers}{logical: scout the number of blockers on each attack?}

\item{transition_sets}{logical: scout sets in transition? If \code{FALSE}, just the endpoint of each attack (i.e. the dig) and the subsequent counter-attack are scouted}
}
\value{
A named list
}
\description{
Scouting options
}
