% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_scouter.R
\name{ov_scouting_options}
\alias{ov_scouting_options}
\title{Scouting options}
\usage{
ov_scouting_options(
  end_convention = "actual",
  nblockers = TRUE,
  default_nblockers = NA,
  transition_sets = FALSE,
  attacks_by = "codes",
  team_system = "SHM3",
  setter_dump_code = "PP",
  second_ball_attack_code = "P2",
  overpass_attack_code = "PR",
  scout_name = "",
  show_courtref = FALSE
)
}
\arguments{
\item{end_convention}{string: either "actual" or "intended". Is the end coordinate of an attack or serve the actual end location (where the ball contacted the floor or out of bounds area), or the intended one. The actual might differ from the intended if there is a block touch or the ball hit the net. If "actual", and a block touch is recorded, then the end location of the attack will not be used for the dig location (the dig location will be missing)}

\item{nblockers}{logical: scout the number of blockers on each attack?}

\item{default_nblockers}{integer: if \code{nblockers} is TRUE, what number of blockers should we default to? If \code{NA}, no default}

\item{transition_sets}{logical: scout sets in transition? If \code{FALSE}, just the endpoint of each attack (i.e. the dig) and the subsequent counter-attack are scouted}

\item{attacks_by}{string: "codes" (X5, V5, etc) or "tempo" (high, medium, quick)}

\item{team_system}{string: the assumed system that teams are using to assign e.g. passing and hitting responsibilities
\itemize{
\item "SHM3" - a setter-hitter-middle rotation, with 3 passers (the libero and two outside hitters)
}}

\item{setter_dump_code}{string: the attack combination code for a setter dump}

\item{second_ball_attack_code}{string: the attack combination code for a second-ball attack}

\item{overpass_attack_code}{string: the attack combination code for an attack on an overpass}

\item{scout_name}{string: the name of the scout (your name)}

\item{show_courtref}{logical: if \code{TRUE}, show the court reference lines overlaid on the video}
}
\value{
A named list
}
\description{
Scouting options
}
